Manual import

+++ upload/admin/controller/module/exchange1c.php
@@ -12,14 +12,12 @@ class ControllerModuleExchange1c extends Controller {
 		$this->load->model('setting/setting');
 			
 		if (($this->request->server['REQUEST_METHOD'] == 'POST') && $this->validate()) {
-			$this->model_setting_setting->editSetting('exchange1c', $this->request->post);		
-					
+			$this->model_setting_setting->editSetting('exchange1c', $this->request->post);				
 			$this->session->data['success'] = $this->language->get('text_success');
-						
 			$this->redirect($this->url->link('extension/module', 'token=' . $this->session->data['token'], 'SSL'));
 		}
 
-		$this->data['version'] = 'Version 121220.git';
+		$this->data['version'] = 'Version 121222.git';
 
 		$this->data['heading_title'] = $this->language->get('heading_title');
 		$this->data['entry_username'] = $this->language->get('entry_username');
@@ -29,6 +27,7 @@ class ControllerModuleExchange1c extends Controller {
 		$this->data['entry_flush_category'] = $this->language->get('entry_flush_category');
 		$this->data['entry_flush_manufacturer'] = $this->language->get('entry_flush_manufacturer');
 		$this->data['entry_flush_quantity'] = $this->language->get('entry_flush_quantity');
+		$this->data['entry_flush_attribute'] = $this->language->get('entry_flush_attribute');
 		$this->data['entry_fill_parent_cats'] = $this->language->get('entry_fill_parent_cats');
 		$this->data['entry_seo_url'] = $this->language->get('entry_seo_url');
 		$this->data['text_yes'] = $this->language->get('text_yes');
@@ -38,13 +37,17 @@ class ControllerModuleExchange1c extends Controller {
 		$this->data['text_tab_general'] = $this->language->get('text_tab_general');
 		$this->data['text_tab_product'] = $this->language->get('text_tab_product');
 		$this->data['text_tab_order'] = $this->language->get('text_tab_order');
+		$this->data['text_tab_manual'] = $this->language->get('text_tab_manual');
 		$this->data['text_empty'] = $this->language->get('text_empty');
+		$this->data['text_max_filesize'] = sprintf($this->language->get('text_max_filesize'), @ini_get('max_file_uploads'));
 		$this->data['text_homepage'] = $this->language->get('text_homepage');
 		$this->data['entry_status'] = $this->language->get('entry_status');
 		$this->data['entry_order_status'] = $this->language->get('entry_order_status');
 		$this->data['entry_use_utf8'] = $this->language->get('entry_use_utf8');
 		$this->data['entry_notify'] = $this->language->get('entry_notify');
 		$this->data['entry_fill_parent_cats'] = $this->language->get('entry_fill_parent_cats');
+		$this->data['entry_upload'] = $this->language->get('entry_upload');
+		$this->data['button_upload'] = $this->language->get('button_upload');
 
 		$this->data['button_save'] = $this->language->get('button_save');
 		$this->data['button_cancel'] = $this->language->get('button_cancel');
@@ -87,13 +90,14 @@ class ControllerModuleExchange1c extends Controller {
       		'separator' => ' :: '
    		);
 
+   		$this->data['token'] = $this->session->data['token'];
 		
 		//$this->data['action'] = HTTPS_SERVER . 'index.php?route=module/exchange1c&token=' . $this->session->data['token'];
 		$this->data['action'] = $this->url->link('module/exchange1c', 'token=' . $this->session->data['token'], 'SSL');
 
 		//$this->data['cancel'] = HTTPS_SERVER . 'index.php?route=extension/exchange1c&token=' . $this->session->data['token'];
 		$this->data['cancel'] = $this->url->link('extension/module', 'token=' . $this->session->data['token'], 'SSL');
-		
+
 		if (isset($this->request->post['exchange1c_username'])) {
 			$this->data['exchange1c_username'] = $this->request->post['exchange1c_username'];
 		} else {
@@ -140,6 +144,12 @@ class ControllerModuleExchange1c extends Controller {
 			$this->data['exchange1c_flush_quantity'] = $this->request->post['exchange1c_flush_quantity'];
 		} else {
 			$this->data['exchange1c_flush_quantity'] = $this->config->get('exchange1c_flush_quantity');
+		}
+
+		if (isset($this->request->post['exchange1c_flush_attribute'])) {
+			$this->data['exchange1c_flush_attribute'] = $this->request->post['exchange1c_flush_attribute'];
+		} else {
+			$this->data['exchange1c_flush_attribute'] = $this->config->get('exchange1c_flush_attribute');
 		}
 
 		if (isset($this->request->post['exchange1c_fill_parent_cats'])) {
@@ -193,6 +203,7 @@ class ControllerModuleExchange1c extends Controller {
 	}
 
 	private function validate() {
+
 		if (!$this->user->hasPermission('modify', 'module/exchange1c')) {
 			$this->error['warning'] = $this->language->get('error_permission');
 		}
@@ -201,7 +212,7 @@ class ControllerModuleExchange1c extends Controller {
 			return TRUE;
 		} else {
 			return FALSE;
-		}	
+		}
 	}
 	
 	public function install() {}
@@ -230,12 +241,12 @@ class ControllerModuleExchange1c extends Controller {
 		}
 		
 		// Авторизуем
-		if(($this->config->get('exchange1c_username') != '') && ($_SERVER['PHP_AUTH_USER'] != $this->config->get('exchange1c_username'))) {
+		if(($this->config->get('exchange1c_username') != '') && (@$_SERVER['PHP_AUTH_USER'] != $this->config->get('exchange1c_username'))) {
 			echo "failure\n";
 			echo "error login";
 		}
 		
-		if(($this->config->get('exchange1c_password') != '') && ($_SERVER['PHP_AUTH_PW'] != $this->config->get('exchange1c_password'))) {
+		if(($this->config->get('exchange1c_password') != '') && (@$_SERVER['PHP_AUTH_PW'] != $this->config->get('exchange1c_password'))) {
 			echo "failure\n";
 			echo "error password";
 			exit;
@@ -245,8 +256,66 @@ class ControllerModuleExchange1c extends Controller {
 		echo session_name()."\n";
 		echo session_id() ."\n";
 	}
+
+	public function manualImport() {
+		$this->load->language('module/exchange1c');
+
+		$cache = DIR_CACHE . 'exchange1c/';
+		$json = array();
+
+		if (!empty($this->request->files['file']['name'])) {
+
+			$this->modeCatalogInit(false);
+			$zip = new ZipArchive;
+			
+			if ($zip->open($this->request->files['file']['tmp_name']) === true) {
+				$zip->extractTo($cache);
+
+				if (file_exists($cache . 'import.xml')) {
+					$this->modeImport('import.xml');
+				}
+
+				if (file_exists($cache . 'offers.xml')) {
+					$this->modeImport('offers.xml');
+				}
+
+				if (is_dir($cache . 'import_files')) {
+					$images = DIR_IMAGE . 'import_files/';
+					
+					if (is_dir($images)) {
+						$this->cleanDir($images);
+					}
+
+					rename($cache . 'import_files/', $images);
+				}
+
+			} else {
+
+				$handle = fopen($this->request->files['file']['tmp_name'], 'r');
+				$buffer = fread($handle, 256);
+				fclose($handle);
+
+				if (strpos($buffer, 'Классификатор')) {
+					move_uploaded_file($this->request->files['file']['tmp_name'], $cache . 'import.xml');
+					$this->modeImport('import.xml');
+				
+				} else if (strpos($buffer, 'ПакетПредложений')) {
+					move_uploaded_file($this->request->files['file']['tmp_name'], $cache . 'offers.xml');
+					$this->modeImport('offers.xml');
+
+				} else {
+					$json['error'] = $this->language->get('text_upload_error');
+					exit;
+				}
+			}
+
+			$json['success'] = $this->language->get('text_upload_success');
+		}
+
+		$this->response->setOutput(json_encode($json));
+	}
 	
-	public function modeCatalogInit() {
+	public function modeCatalogInit($echo = true) {
 		
 		$this->load->model('tool/exchange1c');
 		
@@ -257,7 +326,7 @@ class ControllerModuleExchange1c extends Controller {
 		$this->model_tool_exchange1c->checkDbSheme();
 		
 		// Удаляем товары
-		if($this->config->get('exchange1c_flush_product')) {
+		if ($this->config->get('exchange1c_flush_product')) {
 			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'product');
 			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'product_attribute');
 			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'product_description');
@@ -277,7 +346,7 @@ class ControllerModuleExchange1c extends Controller {
 		}
 
 		// Очищает таблицы категорий
-		if($this->config->get('exchange1c_flush_category')) {
+		if ($this->config->get('exchange1c_flush_category')) {
 			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'category'); 
 			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'category_description');
 			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'category_to_store');
@@ -287,7 +356,7 @@ class ControllerModuleExchange1c extends Controller {
 		}
 			
 		// Очищает таблицы от всех производителей
-		if($this->config->get('exchange1c_flush_manufacturer')) {
+		if ($this->config->get('exchange1c_flush_manufacturer')) {
 			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'manufacturer');
 			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'manufacturer_description');
 			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'manufacturer_to_store');
@@ -295,11 +364,13 @@ class ControllerModuleExchange1c extends Controller {
 		}
 			
 		// Очищает атрибуты
-		$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'attribute');
-		$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'attribute_description');
-		$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'attribute_to_1c');
-		$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'attribute_group');
-		$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'attribute_group_description');
+		if ($this->config->get('exchange1c_flush_attribute')) {
+			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'attribute');
+			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'attribute_description');
+			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'attribute_to_1c');
+			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'attribute_group');
+			$this->db->query('TRUNCATE TABLE ' . DB_PREFIX . 'attribute_group_description');
+		}
 
         
 		// Выставляем кол-во товаров в 0
@@ -309,8 +380,10 @@ class ControllerModuleExchange1c extends Controller {
 
 		$limit = 100000 * 1024;
 	
-		echo "zip=no\n";
-		echo "file_limit=".$limit."\n";
+		if ($echo) {
+			echo "zip=no\n";
+			echo "file_limit=".$limit."\n";
+		}
 	
 	}
 
@@ -326,7 +399,7 @@ class ControllerModuleExchange1c extends Controller {
 		$cache = DIR_CACHE . 'exchange1c/';
 		
 		// Проверяем на наличие имени файла
-		if( isset($this->request->get['filename']) ) {
+		if (isset($this->request->get['filename'])) {
 			$uplod_file = $cache . $this->request->get['filename'];
 		} else {
 			echo "failure\n";
@@ -372,22 +445,25 @@ class ControllerModuleExchange1c extends Controller {
 	
 	}
 	
-	public function modeImport() {
+	public function modeImport($manual = false) {
 		
 		$cache = DIR_CACHE . 'exchange1c/';
-	
-		// Проверяем на наличие имени файла
-		if( isset($this->request->get['filename'])) {
-			$importFile = $cache . $this->request->get['filename'];
+
+		if ($manual) {
+			$filename = $manual;
+			$importFile = $cache . $filename;
+		} else if (isset($this->request->get['filename'])) {
+			$filename = $this->request->get['filename'];
+			$importFile = $cache . $filename;
 		} else {
 			echo "failure\n";
-			echo "ERROR 10: No file name variable" . $this->request->get['filename'];
+			echo "ERROR 10: No file name variable";
 			return 0;
 		}
 		
 		$this->load->model('tool/exchange1c');
 		
-		if($this->request->get['filename'] == 'import.xml') {
+		if($filename == 'import.xml') {
 			
 			$this->model_tool_exchange1c->parseImport();
 
@@ -397,26 +473,31 @@ class ControllerModuleExchange1c extends Controller {
 
 			if ($this->config->get('exchange1c_seo_url')) {
 				$this->load->model('module/deadcow_seo');
-				$this->model_module_deadcow_seo->generateCategories('[category_name]', 'Russian');
-				$this->model_module_deadcow_seo->generateProducts('[product_name]', 'Russian');
+				$this->model_module_deadcow_seo->generateCategories($this->config->get('deadcow_seo_categories_template'), 'Russian');
+				$this->model_module_deadcow_seo->generateProducts($this->config->get('deadcow_seo_products_template'), 'Russian');
+				$this->model_module_deadcow_seo->generateManufacturers($this->config->get('deadcow_seo_manufacturers_template'), 'Russian');
 			}
 
-			echo "success\n";
+			if (!$manual) {
+				echo "success\n";
+			}
 			
-		} elseif($this->request->get['filename'] == 'offers.xml') {
+		} else if ($filename == 'offers.xml') {
 			
 			$this->model_tool_exchange1c->parseOffers();
-			echo "success\n";
+			
+			if (!$manual) {
+				echo "success\n";
+			}
 			
 		} else {
 		
 			echo "failure\n";
-			echo $this->request->get['filename'];
+			echo $filename;
 			
 		}
 		
 		$this->cache->delete('product');
-		
 		return;
 	}
